+++ utility.c
@@ -74,7 +74,7 @@ char* get_rss_virt_mem(void) {
 }
 
 
-void exitWithDealloc(bool error, unsigned int num_vertex, FILE * fp_dag, row_g *rows, pthread_t *threads, t_args *args, pthread_mutex_t *roots_mutex, int *roots, row_l *labels, FILE *fp_query, el_query *queries){
+void exitWithDealloc(bool error, unsigned long num_vertex, FILE * fp_dag, row_g *rows, pthread_t *threads, t_args *args, pthread_mutex_t *roots_mutex, unsigned long *roots, row_l *labels, FILE *fp_query, el_query *queries){
     //1. error in: rows = (row_g *) malloc (num_vertex * sizeof (row_g));
     if(fp_dag)
         fclose(fp_dag);
@@ -142,15 +142,15 @@ void exitWithDealloc(bool error, unsigned int num_vertex, FILE * fp_dag, row_g *
 }
 
 //Swap e randomize are utility functions used for randomize the roots
-void swap (int *a, int *b) {
-    int temp = *a;
+void swap (unsigned long *a, unsigned long *b) {
+    unsigned long temp = *a;
     *a = *b;
     *b = temp;
 }
 
-void randomize(int *array, int n) {
-    for(int i=n-1; i>0; i--) {
-        int j = rand() % (i+1);
+void randomize(unsigned long *array, unsigned long n) {
+    for(unsigned long i=n-1; i>0; i--) {
+        unsigned long j = rand() % (i+1);
         if (i != j)
             swap(&array[i], &array[j]);
     }
